backend/requirements.txt
fastapi
uvicorn
python-dotenv
numpy
deepgram-sdk
groq